requirements.txt
pandas
numpy
pyyaml
dvc[s3]
scikit-learn
flask 
flask-cors
python-dotenv
xmltodict